Move Youtube definitions to their own files
