+++ services/graphql-server/src/graphql/definitions/google-data-api/index.js
@@ -1,0 +1,8 @@
+const gql = require('graphql-tag');
+const youtube = require('./youtube');
+
+module.exports = gql`
+
+${youtube}
+
+`;
