+++ services/graphql-server/src/graphql/definitions/platform/content/types/company.js
@@ -35,50 +35,6 @@ type ContentCompanyYoutube {
   playlistId: String
 }
 
-type YoutubePlaylistItemsApiResponse {
-  pageInfo: YoutubePageInfo
-  items: [YoutubePlaylistItem!]!
-  error: String
-}
-
-type YoutubePageInfo {
-  totalResults: Int
-  resultsPerPage: Int
-}
-
-type YoutubePlaylistItem {
-  id: String
-  snippet: YoutubePlaylistItemSnippet
-}
-
-type YoutubePlaylistItemSnippet {
-  publishedAt: Date
-  channelId: String
-  title: String
-  description: String
-  thumbnails: YoutubeVideoThumbnails
-  channelTitle: String
-  playlistId: String
-  position: Int
-  resourceId: YoutubePlaylistItemResourceId
-}
-
-type YoutubePlaylistItemResourceId {
-  videoId: String
-}
-
-type YoutubeVideoThumbnails {
-  default: YoutubeVideoThumbnail
-  medium: YoutubeVideoThumbnail
-  high: YoutubeVideoThumbnail
-}
-
-type YoutubeVideoThumbnail {
-  url: String
-  width: Int
-  height: Int
-}
-
 type ContentCompanyConnection {
   totalCount: Int!
   edges: [ContentCompanyEdge]!
